Dep.Bricks.Box

Description

This module defines characters to generate images in the user interface. These render wires, boxes, etc.

Box borders

boxh :: Char

A character to render the horizontal border corner of a thick box.

boxv :: Char

A character to render the vertical border corner of a thick box.

boxlb :: Char

A character to render the bottom left corner of a thick box.

boxlt :: Char

A character to render the top left corner of a thick box.

boxrb :: Char

A character to render the bottom right corner of a thick box.

boxrt :: Char

A character to render the top right corner of a thick box.

boxhu :: Char

A character to render the horizontal border of a thick box with a wire at at the upper part.

boxhd :: Char

A character to render the horizontal border of a thick box with a wire at at the bottom part.

boxvl :: Char

A character to render the vertical border of a thick box with a wire at at the left part.

boxvr :: Char

A character to render the vertical border of a thick box with a wire at at the right part.

Wire characters

lineh :: Char

A character to render a wire in a horizontal direction.

linev :: Char

A character to render a wire in a vertical direction.

negator :: Char

A character to render a negator.