Dep.Bricks.Gate

Description

This module has functions to create gates with an arbitrary number of inputs.

Create generic gates

gateH

Arguments

:: Char
The Char that will be printed (repeatedly) on the gate to specify the type of gate.

-> Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

Create a gate in the horizontal direction with the given Char that specifies what type of gate it is, and the number of input wires.

gateV

Arguments

:: Char
The Char that will be printed (repeatedly) on the gate to specify the type of gate.

-> Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

Create a gate in the vertical direction with the given Char that specifies what type of gate it is, and the number of input wires.

gate

Arguments

:: Char
The Char that will be printed (repeatedly) on the gate to specify the type of gate.

-> CircuitLayout
The given CircuitLayout that specifies in what direction the gate is written.

-> Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

Create a gate in the given CircuitLayout with the given Char that specifies what type of gate it is, and the number of input wires.

genericGate

Arguments

:: Char
The Char that specifies what type of gate it is. This Character will be printed (repeatedly) on the gate.

-> CircuitLayout
The CircuitLayout that specifies in what direction the gate is written.

-> [Bool]
Specifies what wires have negators at the input side.

-> Bool
Specifies if the output side has a negator.

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A function to create a gate with negators and the input and output.

A collection of and gates

andGate

Arguments

:: CircuitLayout
The given CircuitLayout that specifies in what direction the gate is written.

-> Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create and gates in a more effective way.

andGateH

Arguments

:: Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create horizontal and gates in a more effective way.

andGateH2

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a horizontal and gate with two input wires.

andGateH3

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a horizontal and gate with three input wires.

andGateV

Arguments

:: Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create vertical and gates in a more effective way.

andGateV2

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a vertical and gate with two input wires.

andGateV3

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a vertical and gate with three input wires.

A collection of or gates

orGate

Arguments

:: CircuitLayout
The given CircuitLayout that specifies in what direction the gate is written.

-> Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create or gates in a more effective way.

orGateH

Arguments

:: Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create horizontal or gates in a more effective way.

orGateH2

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a horizontal or gate with two input wires.

orGateH3

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a horizontal or gate with three input wires.

orGateV

Arguments

:: Int
The number of input wires (fan-in).

-> Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create vertical or gates in a more effective way.

orGateV2

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a vertical or gate with two input wires.

orGateV3

Arguments

:: Attr
The Attr that specifies the style of the gate.

-> Image
The corresponding Image.

A helper function to create a vertical or gate with three input wires.