Dep.Bricks.Karnaugh

Description

A module to render, update, and interact with Karnaugh cards.

Documentation

renderKarnaugh

Arguments

:: CharRenderable a

=> Three a
The given Three to render as a Karnaugh card.

-> SumOfProducts
The sum of products that will be used to mark the Karnaugh card.

-> [String]
The names of the variables that are rendered. If there are no sufficient variables, it will work with x₀, x₁, x₂, etc.

-> Attr
The base Attribute to render the Karnaugh card.

-> Image
The image that contains a rendered version of the Karnaugh card.

Render the given Three as a Karnaugh card, that can also visualize the sum-of-product.

renderKarnaugh'

Arguments

:: Three ThreeValue
The given Three to render as a Karnaugh card.

-> [String]
The names of the variables that are rendered. If there are no sufficient variables, it will work with x₀, x₁, x₂, etc.

-> Attr
The base Attribute to render the Karnaugh card.

-> Image
The image that contains a rendered version of the Karnaugh card.

Render the given Three as a Karnaugh card, that can also visualize the sum-of-product.