Dep.Bricks.Layered

Description

This module exports the mergeLayers function that combine multiple Layers in a single layer. This is used if writing several layers is easier than writing a single image.

A type synoniem for one layer.

type Layer a = [[a]]

A raster is a list of lists of characters.

Merge several layers to one layer.

createImage

Arguments

:: [[(Attr, Char)]]
The given list of lists of Chars

-> Image

Convert a list of lists of 2-tuples that contain an Attributes and a Char to an Image.

mergeLayers

Arguments

:: (Char -> Bool)
The predicate that determines if the upper layer will override the layer directly below.

-> [(Attr, Layer Char)]
The list of Layers, the first item is the top item and the last item is the bottom layer.

-> Image
An image that we create by merging the different layers together.

Merge the given Layers with the corresponding Attributes to a single Image.