Dep.Bricks.Layout

Description

Specifies how a circuit is rendered. This can be in a horizontal and vertical manner.

The orientation of a circuit

data CircuitLayout

A data type that specifies whether the circuit is render in a horizontal or vertical manner.

Constructors

Horizontal
The horizontal orientation.

Vertical
The vertical orientation.