Dep.Bricks.Negation

Description

A module to create images with negations for gates, wires, and other modules.

Negation for a single wire

negation

Arguments

:: Char
The Char to use in case we do not negate.

-> Bool
A Bool that specifies if this is a negator.

-> Attr
The Attr to use when rendering the image.

-> Image
The corresponding Image for the given (optional) negator.

Create an Image that contains a single character that is either the negator or the given Char.

negationH

Arguments

:: Bool
Specifies if this is a simple wire (False), or a negator (True).

-> Attr
The Attr to use when rendering the image.

-> Image
The corresponding Image for the given (optional) negator.

Create an image for a negator on a horizontal wire.

negationV

Arguments

:: Bool
Specifies if this is a simple wire (False), or a negator (True).

-> Attr
The Attr to use when rendering the image.

-> Image
The corresponding Image for the given (optional) negator.

Create an image for a negator on a vertical wire.

Negation for a list of wires

negationList

Arguments

:: (Image -> Image -> Image)
A function to merge two Images together.

-> Char
A Char that specifies what to render in case we do not render a negator.

-> [Bool]
The list of Bools that specifies when a negator occurs.

-> Attr
The Attr that specifies the style of the negators.

-> Image
The corresponding Image that contains wires and negators.

Create a list of negators in one of the two directions. The Char specifies what to render in case it is not a negator.

negationHList

Arguments

:: [Bool]
The list of Bools that specifies when a negator occurs.

-> Attr
The Attr that specifies the style of the negators.

-> Image
The corresponding Image that contains wires and negators.

Create an Image of negators in the horizontal direction. The wires here are in the vertical direction.

negationVList

Arguments

:: [Bool]
The list of Bools that specifies when a negator occurs.

-> Attr
The Attr that specifies the style of the negators.

-> Image
The corresponding Image that contains wires and negators.

Create an Image of negators in the vertical direction. The wires here are in the horizontal direction.