Dep.Bricks.Palette

Description

This module generates palettes. A Palette is an infinite list of Colors that endlessly repeats itself. This is used to obtain colors to typeset different items on the screen.

Palettes

type Palette = [Color]

A palette is an (endless) list of Colors.

isoColorPalette

Arguments

:: Palette
A palette that endlessly repeats itself and exhaustively enumerates the possible ISO colors.

Work with a color palette of the six ISO colors.

brightIsoColorPalette

Arguments

:: Palette
A palette that endlessly repeats itself and exhaustively enumerates the possible ISO colors.

Work with a color palette of the six ISO colors.

color240Palette

Arguments

:: Palette
A palette that endlessly repeats itself and exhaustively enumerates the possible colors 240s.

Work with a color palette with 240 colors, but only 238 are selected, since black and white are not considered to be colors for the palette.

Determine other color for ISOColors

swapIsoColorBright

Arguments

:: Color
The given Color to convert to a more/less bright color.

-> Color
The corresponding variant that is more/less bright of the given Color.

Swap between the bright mode and the normal mode of the given Color.

isoColorToBright

Arguments

:: Color
The given Color to make bright.

-> Color
The corresponding bright variant of the given Color.

Create the brighter color

isoColorFromBright

Arguments

:: Color
The given Color to make a Color that is less bright.

-> Color
The corresponding less bright variant of the given Color.

Obtain the equivalent color that is not bright.