Dep.Class.Opposite

Description

This module defines a typeclass Opposite that specifies that each value of a specific type has an opposite value of the same type.

A typeclass to determine the opposite value.

class Opposite a where

A typeclass where the values of its members have an opposite element from the same type.

Minimal complete definition

Nothing

Methods

opposite

Arguments

:: a
The given item to determine the opposite from.

-> a
The opposite of the given value.

A function that determines the opposite value.

default opposite :: (Functor f, Opposite b, a ~ f b) => a -> a