Dep.Class.Renderable

Description

A module that has a type class to render items of the types that are members of the typeclass.

Documentation

class CharRenderable a where

A typeclass that specifies that a certain type can be rendered with a single Character.

Methods

charRenderItem

Arguments

:: a
THe given item to render.

-> Char
The corresponding Character when rendering the object.

Render the given item to a Char object.

class Renderable a where

A typeclass that specifies that a certain type can be rendered as a DEP Text object.

Minimal complete definition

Nothing

Methods

renderItem

Arguments

:: a
The given item to render.

-> Text
The corresponding Text for the rendered information.

Render the given item to a Text object.

default renderItem :: CharRenderable a => a -> Text