Dep.Class.Simplify

Description

The Simplify typeclass exports a function simplify that is used to simplify a value. That function should be idempotent: multiple simplify calls should not have a different result than a single simplify call.

A typeclass to specify that values of the type can be simplified.

class Simplify a where

A typeclass that specifies that the the values of the given type sometimes can be simplfied. The simplify operation is idempotent: calling it multiple times has the same effect as calling it once.

Methods

simplify

Arguments

:: a
The given object to simplify.

-> a
The correspoding simplified relevant.

A function to simplify the given object to a potentially simplier object. This function should be idempotent: calling simplify multiple times has the same effect as calling simplify once.