Dep.Class.Walkable

Description

This module defines a typeclass Walkable that specifies that one can walk through a deterministic data structure for a given type of "steps".

Deterministic steps and walks

class Walkable f step | f -> step where

A typeclass that specifies that we can walk through a datastructure with steps.

Minimal complete definition

step | walk

Methods

step

Arguments

:: f a
The original object where we will make a step.

-> step
The given step we take on the given data structure.

-> f a
The result of an object when we take one step.

Take one step with the given step parameter, and return an object with the same type.

stepValues'

Arguments

:: Foldable f

=> f a
The initial state.

-> step
The step that we take to move to another state.

-> [a]
A list of tail elements that can be added to the result.

-> [a]
The list of values from the modified state together with the given tail.

Obtain the values of the item after making a step.

stepValues

Arguments

:: Foldable f

=> f a
The initial state.

-> step
The step that we take to move to another state.

-> [a]
The list of values from the modified state.

Obtain the values of the item after making a step.

allStep

Arguments

:: Functor g

=> g (f a)
The collection of items on which we apply a step.

-> step
The given step that will be applied.

-> g (f a)
A collection of items that are the result of making a step for each input item.

Apply the same step for all the items in the given collection (functor) of items.

allStepValues'

Arguments

:: (Foldable f, Foldable g)

=> g (f a)
The initial states.

-> step
The step that we take to move to another state.

-> [a]
A list of tail elements that can be added to the result.

-> [a]
The list of values from the modified state together with the given tail.

Obtain the values of the items after making the same step for all items.

allStepValues

Arguments

:: (Foldable f, Foldable g)

=> g (f a)
The initial states.

-> step
The step that we take to move to another state.

-> [a]
The list of values from the modified state.

Obtain the values of the items after making the same step for all items.

walk

Arguments

:: Foldable g

=> f a
The original object where we will make a step.

-> g step
The given foldable of steps we take on the given data structure.

-> f a
The result of an object when we take one step.

Take a sequence of steps with the given list of steps.

walkValues'

Arguments

:: (Foldable f, Foldable g)

=> f a
The initial state for which we will make a walk.

-> g step
A sequene of steps that describe the walk.

-> [a]
A list of tail elements that can be added at the end of the list.

-> [a]
The values found after making the a walk with the given Walkable in the Foldable.

Take a sequence of steps with the given list of steps and return the values out of the target state.

allWalkValues

Arguments

:: (Foldable f, Foldable g, Functor g, Foldable h)

=> g (f a)
The given collection of initial states over which a fold is made.

-> h step
A Foldable of steps that we will take for all states.

-> [a]
The corresponding values that are wrapped in the target states.

Obtain all the values the Foldable f spans after performing the same (Foldable) of steps on all the given intial states.

allWalk

Arguments

:: (Functor g, Foldable h)

=> g (f a)
The given collection of initial objects.

-> h step
The given foldable of steps that we take.

-> g (f a)
The result collection of items after applying the steps.