Dep.Data.Product

Description

This module provides utility functions to compress/decompress products, and render these Product's to a unicode string.

Type synonyms to represent synthesis

newtype Product

A data type that can be used to specify a product. By using a newtype, we can add special instance to the Product.

Constructors

Product Product'

type Product' = ThreePath

A type alias for a product that is a ThreePath.

newtype CompactProduct

A data type that can be used to specify a CompactProduct. By using a new type, this means that we can define other instances than these for a list of Int.

Constructors

CompactProduct CompactProduct'

type CompactProduct' = [Int]

A more compact representation of a product where the indexes that have Zero or One are listed by the positive or negative index respectively.

newtype SumOfProducts

A data type that is used to specify a sum of products. This type can be used to specify new instance other than these of a list of lists of Ints.

Constructors

SumOfProducts [Product]

type SumOfProducts' = [Product']

A type synonym to present a sum of products where each item of the list is a Product'.

Print products and sums-of-products

showSumOfProducts

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> SumOfProducts'
The given sum of products to convert to a Text.

-> Text
The corresponding Text object that presents the given SumOfProducts'.

Convert the given sum of products to a Text object that presents the SumOfProducts' as a Text object with variables as subscript.

>>> showSumOfProducts 'x' [[One, One], [DontCare, Zero, One]]
"x₀x₁ + x₁'x₂"

showProduct

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> Product'
The given product to convert to a Text.

-> Text
The corresponding Text object that presents the given Product'.

Print a given product as a sequence of variables that can be negated. for example:

>>> showProduct 'x' [One, DontCare, Zero, One]
"x₀x₂'x₃"

showProduct'

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> Text
The text that will be added as tail, this is useful if we combine products.

-> Product'
The given product to convert to a Text.

-> Text
The corresponding Text object that presents the given Product'.

Print a given product as a sequence of variables that can be negated. for example:

>>> showProduct' 'x' mempty [One, DontCare, Zero, One]
"x₀x₂'x₃"

subscriptVariable

Arguments

:: Char
The "root" variable that we will use with a subscript.

-> Int
The subscript that will be added to the character.

-> Text
A text object that contains the "root" variable with the subscript.

Convert the given number to a Text object containing the "root" variable name and the index as subscript.