Dep.Data.Sum

Description

This module provides utility functions to compress/decompress sums, and render these Product's to a unicode string.

Type synonyms to represent synthesis

newtype Sum

A data type that can be used to specify a sum. By using a newtype, we can add special instance to the Sum.

Constructors

Sum Sum'

type Sum' = ThreePath

A type alias for a sum that is a ThreePath.

newtype CompactSum

A data type that can be used to specify a CompactSum. By using a new type, this means that we can define other instances than these for a list of Int.

Constructors

CompactSum CompactSum'

type CompactSum' = [Int]

A more compact representation of a sum where the indexes that have Zero or One are listed by the positive or negative index respectively.

newtype ProductOfSums

A data type that is used to specify a product of sums. This type can be used to specify new instance other than these of a list of lists of Ints.

Constructors

ProductOfSums [Sum]

type ProductOfSums' = [Sum']

A type synonym to present a product of sums where each item of the list is a Sum'.

Print sums and product of sums

showProductOfSums

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> ProductOfSums'
The given sum of products to convert to a Text.

-> Text
The corresponding Text object that presents the given ProductOfSums'.

Convert the given product of sums to a Text object that presents the ProductOfSums' as a Text object with variables as subscript.

>>> showProductOfSums 'x' [[One, One], [DontCare, Zero, One]]
"(x₀x₁)(x₁'x₂)"

showSum

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> Sum'
The given sum to convert to a Text.

-> Text
The corresponding Text object that presents the given Product'.

Print a given sum as a sequence of variables that can be negated. for example:

>>> showSum 'x' [One, DontCare, Zero, One]
"x₀ + x₂' + x₃"

showSum'

Arguments

:: Char
The name of the root variable that will be used with subscripts.

-> Text
The text that will be added as tail, this is useful if we combine Sums.

-> Sum'
The given sum to convert to a Text.

-> Text
The corresponding Text object that presents the given Sum'.

Print a given sum as a sequence of variables that can be negated. for example:

>>> showSum' 'x' mempty [One, DontCare, Zero, One]
"x₀ + x₂' + x₃"