Dep.Data.Three

Description

This modules defines a three, a tree-like data structure with a leaf, a node, and a link where both subtrees are the same. This is used to make more compact and efficient representations of a boolean function.

Defining a three

data Three a

A data structure used to specify a mapping from a list of booleans to a value in a more compact way. This datastructure can effectively be used to define a sum of products or a product of sums.

Constructors

Leaf a
A leaf that contains a single value.

Link (Three a)
A link where it means that this variable does not matter but the next one(s) will.

Split (Three a) (Three a)
A split where this variable determines the outcome.

Paths in a three

type ThreePath = [ThreeStep]

A type of a list (non-deterministic) steps in a Three structure.

type ThreeStep = ThreeValue

A type alias for (non-deterministic) steps in a Three structure.

Catamorphisms

three :: (a -> b) -> (b -> b) -> (b -> b -> b) -> Three a -> b

A catamorphism on the Three object. Here one can provide functions to use for the Leaf, the Link and the Split to catamorph a Three object to another object.

depth

Arguments

:: Three a
The Three object to determine the maximum depth from.

-> Int
The depth of the given Three object.

Determine the maximum depth of the Three tree.

leftmost

Arguments

:: Three a
The Three to obtain the leftmost item from.

-> a
The leftmost item in the given Three.

Obtain the leftmost item of the Three.

rightmost

Arguments

:: Three a
The Three to obtain the rightmost item from.

-> a
The rightmost item in the given Three.

Obtain the rightmost item of the Three.

Manipulate a Three

flipThree

Arguments

:: Three a
The given Three to flip.

-> Three a
The corresponding flipped Three.

Flip the most basic level such that True now maps on False and vice versa. This is for example used to render Karnaugh cards.

flipAllThree

Arguments

:: Three a
The given Three to flip.

-> Three a
The corresponding flipped Three.

Flip all the nodes in the Three such that the False subtree is now a True subtree, and the True subtree is now the False subtree.

Lookups and constructions

nstep

Arguments

:: NonDeterministicWalkable f step

=> f a
The given initial state where we make a non-determinstic step.

-> step
The step that we make, such step can result in zero, one or more new states.

-> [f a]
The list of the possible states with the new step.

Take a non-deterministic step that can result in multiple outcomes.

apply

Arguments

:: (a -> a)
The given function to apply for the items that satisfy the given path.

-> ThreePath
The given path to use.

-> Three (a -> a)
A Three object of functions where the elements that satisfy the path will use the given function and the others will use id.

Construct a Three that will apply the given function for the items that satisfy the given path of three-valued objects.

applyTo

Arguments

:: (a -> a)
The given function to apply to some parts of the Three.

-> ThreePath
The given path that specifies what for what parts of the Three we should apply the function.

-> Three a
The given Three where (part) of the Three will be modified with a given function.

-> Three a
The resulting Three after applying the given function to parts of the Three that satisfy the given path.

Apply the given function to the elements in the given Three that satisfy the given path.

wipe

Arguments

:: a
The given value such that we replace parts of the three with the given ThreePath to this value.

-> ThreePath
The given path that specifies what for what parts of the Three we will set the value.

-> Three a
The given Three where (part) of the Three will be changed to a Leaf with the given value.

-> Three a
The resulting Three after changing the items that match the given ThreePath to the given value.

Convert the given items that match the given ThreePath to the Leaf with a given element.

wipeAll

Arguments

:: a
The given element to use when we wipe.

-> Three a
The original Three where we will wipe items.

-> [ThreePath]
The list of paths to wipe.

-> Three a
The Three object after wiping each element in the given list.

Wipe with the given value all the given ThreePaths.

Retrieve children according to a path

children

Arguments

:: ThreePath
The given ThreePath for the query.

-> Three a
The given Three that we query.

-> [a]
A list of children that satisfy the given ThreePath.

Obtain the children that satisfy a given ThreePath.

children'

Arguments

:: ThreePath
The given ThreePath for the query.

-> Three a
The given Three that we query.

-> [a]
The list of tail elements.

-> [a]
The list of children followed by the given list of tail elements.

Obtain the children that satisfy the given ThreePath.

Convert the Three to an key-value list

toTraces

Arguments

:: Three a

-> [(ThreeValues, a)]
The list of addresses in *reverse* with the corresponding value.

Convert the given Three to a list of 2-tuples with as first item the "address" in reverse, and as second item the value associated with this.

toTraces'

Arguments

:: Three a

-> [(ThreeValues, a)]
The list of trailing items that can be added at the end.

-> [(ThreeValues, a)]
The list of addresses in reverse with the corresponding value.

Convert the given Three to a list of 2-tuples with as first item the "address" in reverse, and as second item the value associated with this.

toTraces''

Arguments

:: Three a

-> ThreeValues
The current address that will be manipulated as we walk through the Three in reverse order.

-> [(ThreeValues, a)]
The list of trailing items that can be added at the end.

-> [(ThreeValues, a)]
The list of addresses in reverse with the corresponding value.

Convert the given Three to a list of 2-tuples with as first item the "address" in reverse, and as second item the value associated with this.