Dep.Data.ThreeValue

Description

A module that defines three-value logic, for example used in thruth tables, Karnaugh cards, etc.

Define three-value logic

data ThreeValue

A data type that is used if a value can present three logical values: don't care (or don't know); zero; and one.

Constructors

Zero
The value is zero or false.

One
The value is one or true.

DontCare
We do not care or do not know the value.

Catamorphisms

threeValue

Arguments

:: a
The value for DontCare.

-> a
The value for Zero.

-> a
The value for One.

-> ThreeValue
The value to convert to one of the given values.

-> a
The corresponding value.

Convert the given ThreeValue object to the corresponding value. This acts as a catamorphism for the ThreeValue type.

toMaybeBool

Arguments

:: ThreeValue
The given ThreeValue to convert.

-> Maybe Bool
The corresponding Bool wrapped in a Maybe.

Convert the given ThreeValue to a Maybe Bool object. where DontCare is mapped to Nothing and Zero and One to True and False wrapped in a Just accordingly.

toChar

Arguments

:: ThreeValue
The given ThreeValue object to map.

-> Char
The corresponding Char that presents the ThreeValue.

Convert the given ThreeValue to a Char that presents the given value. DontCare is mapped to -; Zero to 0; and One to 1.

Convert to and from a ThreeValue

fromBool

Arguments

:: Bool
The given Bool to convert tot a ThreeValue.

-> ThreeValue
The corresponding ThreeValue object.

Convert True and False to Zero and One respectively.

fromMaybeBool

Arguments

:: Maybe Bool
The given Bool object to convert.

-> ThreeValue
The corresponding object where Nothing is mapped to DontCare, and True and False are mapped to One and Zero respectively.

Convert a Maybe Bool to a ThreeValue, where Nothing is mapped to DontCare, and Just True and Just False to One and Zero.

toUpper :: ThreeValue -> Bool

A function that maps Ones and DontCares to Trues; and Zeros to False.

toLower :: ThreeValue -> Bool

A function that maps Zeros and DontCares to Falses; and Ones to True.

Operators on ThreeValue

opposite

Arguments

:: Opposite a

=> a
The given item to determine the opposite from.

-> a
The opposite of the given value.

A function that determines the opposite value.

Type aliasses

type ThreeValues = [ThreeValue]

A type alias for a list of ThreeValue objects.

parseThreeValue :: Stream s m Char => ParsecT s u m ThreeValue

A parser that can parse a single ThreeValue from a Character. The characters for Zero are 0, f or F; the characters for Zero are 1, t or T; and the characters for DontCare are -, d or D,

parseThreeValues :: Stream s m Char => ParsecT s u m ThreeValues

A parser that can parse a (possibly empty) list of ThreeValue from a Character. The characters for Zero are 0, f or F; the characters for Zero are 1, t or T; and the characters for DontCare are -, d or D,

parseThreeValues1 :: Stream s m Char => ParsecT s u m (NonEmpty ThreeValue)

A parser that can parse a NonEmpty list of ThreeValue from a Character. The characters for Zero are 0, f or F; the characters for Zero are 1, t or T; and the characters for DontCare are -, d or D,